Include Shipment Cancelled Callback
The Intelligent Foods Partner API now includes a SHIPMENT_CANCELLED
callback event. This event notifies the client that a shipment is
no longer being sent to a customer, and is in a cancelled state. We
should include this response to allow users to test using a stubbed
cancelled callback.

This change addresses the need by:
* Including a SHIPMENT_CANCELLED callback