Support Multiple Versions of the API
In the future, we will need to integrate with v2 of the IntelligentFoods
API. This migration will take place over some time, and resources will
be migrated slowly, requiring this gem to be able to interact with
different versions of the API on a resource to resource basis. In order
to make that change easy, we should introduce a api adapter factory
which will return the correct api adapter. For this version, only V1 is
required.

This change addresses the need by:
* Introducing a V1 adapter
* Refactoring how the Order and Menu resource determine the base_url
* Moving the tld logic out of the base class and into the adapter
* Moving authentication to the adapter as well